refactored code and added separate file for account management  routes
