+++ server.js
@@ -11,7 +11,6 @@ var Shift = require('./app/models/schedule');
 var port = process.env.PORT || 8080;
 var jwt = require('jwt-simple');
 var SchedFunctions = require('./app/functions/scheduleFun.js');
-
 // get our request parameters
 app.use(bodyParser.urlencoded({
     extended: false
@@ -46,74 +45,9 @@ require('./config/passport')(passport);
 
 // bundle our routes
 var apiRoutes = express.Router();
+var acc = require('./routes/accountRoutes')(apiRoutes);
 
 
-// create a new user account (POST http://localhost:8080/api/signup)
-apiRoutes.options('/signup', cors());
-apiRoutes.post('/signup', function (req, res) {
-    if (!req.body.firstName || !req.body.password) {
-        res.json({
-            success: false,
-            msg: 'Please add name and password.'
-        });
-    } else {
-        var newUser = new User({
-            firstName: req.body.firstName,
-            lastName: req.body.lastName,
-            password: req.body.password,
-            email: req.body.email
-        });
-        // save the user
-        newUser.save(function (err) {
-            if (err) {
-                console.log(err);
-                return res.json({
-                    success: false,
-                    msg: 'Username already exists.'
-                });
-            }
-            res.json({
-                success: true,
-                msg: 'Successful created new user.'
-            });
-        });
-    }
-});
-
-//route to authenticate a user (POST http://localhost:8080/api/authenticate)
-apiRoutes.options('/authenticate', cors());
-apiRoutes.post('/authenticate', function (req, res) {
-    User.findOne({
-        email: req.body.email
-    }, function (err, user) {
-        if (err) console.log(err);
-
-        if (!user) {
-            res.send({
-                success: false,
-                msg: 'Authentication failed. User not found.'
-            });
-        } else {
-            // check if password matches
-            user.comparePassword(req.body.password, function (err, isMatch) {
-                if (isMatch && !err) {
-                    // if user is found and password is right create a token
-                    var token = jwt.encode(user, config.secret);
-                    // return the information including token as JSON
-                    res.json({
-                        success: true,
-                        token: 'JWT ' + token
-                    });
-                } else {
-                    res.send({
-                        success: false,
-                        msg: 'Authentication failed. Wrong password.'
-                    });
-                }
-            });
-        }
-    });
-});
 
 //Schedule routes current*********************************************************************************
 apiRoutes.options('/addAvailability', cors());
@@ -190,41 +124,6 @@ apiRoutes.get('/getAvailability', passport.authenticate('jwt', {
         });
     }
 
-});
-// route to a restricted info (GET http://localhost:8080/api/memberinfo)
-apiRoutes.options('/memberinfo', cors());
-apiRoutes.get('/memberinfo', passport.authenticate('jwt', {
-    session: false
-}), function (req, res) {
-    var token = getToken(req.headers);
-    if (token) {
-        var decoded = jwt.decode(token, config.secret);
-        User.findOne({
-            email: decoded.email
-        }, function (err, user) {
-            if (err) throw err;
-
-            if (!user) {
-                return res.status(403).send({
-                    success: false,
-                    msg: 'Authentication failed. User not found.'
-                });
-            } else {
-                res.json({
-                    success: true,
-                    msg: 'Welcome in the member area ' + user.firstName + '!',
-                    firstName: user.firstName,
-                    lastName: user.lastName,
-                    email: user.email
-                });
-            }
-        });
-    } else {
-        return res.status(403).send({
-            success: false,
-            msg: 'No token provided.'
-        });
-    }
 });
 
 getToken = function (headers) {
